requirements.txt
streamlit
streamlit-webrtc
opencv-python
tensorflow
mysql-connector-python
speechrecognition
pyttsx3
pipwin
pyaudio
scikit-learn
numpy
pandas
bcrypt

